This adds sourdough to the calculator (#36)
I've been experimenting with sourdough for a long time and personally
started to like sourdough based pizza the most.

I added it to the calculator. Depending on the flour you need to reduce
hydration a little bit. Else it could it happen that the dough becomes
too sticky.

We will also need to re-write the main recipe slightly to include this.
Since it's just experimental I will do that once I have done this a few
more times.